Final Submission
Hi Charolette,

Hope you are doing well! I just wanted to give you a heads up, I'm in the social area so the project that I designed is a Qualtrics survey because that's what I'm doing for my Master's. I asked Katherine about this and she said it's okay. I know it's not the most complicated survey out there since one is a correlational study and the other only has two conditions, so I tried to include as much information as I could.

Cheers,
Joe